requirements.txt
Cerberus==0.9.2
Eve==0.6.4
Events==0.2.1
Flask==0.10.1
Flask-PyMongo==0.4.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pymongo==3.4.0
PyYAML==3.12
simplejson==3.10.0
Werkzeug==0.11.3
